Add files via upload
The differential interference synthetic aperture radar (DInSAR) can be used to obtain the interference ring of surface displacement. The denser the interference ring, the greater the change in surface displacement, and the closer to the earthquake-generating fault. The densest place of interference rings in the figure is the northern end of Samos, and only some interference rings are shown, implying that the main coseismic displacement is located on the seafloor near the epicenter, which is also related to small-scale tsunami record.